Move os testes da API do i-Diário para outra pasta
