+++ phpunit.xml
@@ -23,13 +23,12 @@
     </testsuite>
     <testsuite name="Feature">
       <directory suffix="Test.php">./tests/Feature</directory>
-      <exclude>./tests/Feature/DiarioApi</exclude>
     </testsuite>
     <testsuite name="Pages">
         <directory suffix="Test.php">./tests/Pages/</directory>
     </testsuite>
     <testsuite name="Diario">
-      <directory suffix="Test.php">./tests/Feature/DiarioApi</directory>
+      <directory suffix="Test.php">./tests/Api/DiarioApi</directory>
     </testsuite>
   </testsuites>
   <php>
