+++ tests/Api/PromotionFromAverageAndAttendanceWithoutRetakeTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\DiarioApi;
+namespace Tests\Api;
 
 use App\Models\LegacyEnrollment;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
