+++ tests/Api/WithoutScoreDiarioApiTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\DiarioApi;
+namespace Tests\Api;
 
 use App\Models\LegacyEvaluationRule;
 use App_Model_MatriculaSituacao;
